New daemon server w/ NameServer support
LAD (Link Arbiter Daemon) was "stolen" from the CE tree and converted for
use here as a "Link Server Daemon".  Initially it contains support for just
NameServer, along with a front end NameServer_client.c that translates
NameServer APIs into commands/responses to/from the daemon.

This NameServer daemon support contains a remote "get" function that
communicates with remote NameServer databases on other cores using the
socket interface of rpmsg.  It also serves "get" requests from these
other cores.

There is also a NameServerApp.c file which excersizes the NameServer API set,
for the purpose of validating the daemon approach (although the test is mostly
daemon-agnostic, as it would apply to a non-daemon NameServer library).