Only one LAD_connect() per process
Allowing concurrent access to LAD from each thread of a process proved
too costly and complex, so NameServer_client.c was changed to allow
multiple threads to sequentially access LAD.  This support comes in the
form of a mutex that is taken before sending a commang to LAD, and released
after receiving a reply.

MessageQMulti.c was changed to do just one LAD_connect() (instead of one
per thread).